refactor(cli): decompose pipeline errors and reporting helpers
